fixed bug in PHPExcel where null CurrentObject was being written out

+++ PHPExcel/Classes/PHPExcel/CachedObjectStorage/PHPTemp.php
@@ -58,14 +58,16 @@ class PHPExcel_CachedObjectStorage_PHPTemp extends PHPExcel_CachedObjectStorage_
      */
 	protected function _storeData() {
 		if ($this->_currentCellIsDirty) {
-			$this->_currentObject->detach();
+            if ($this->_currentObject) {
+                $this->_currentObject->detach();
 
-			fseek($this->_fileHandle,0,SEEK_END);
-			$offset = ftell($this->_fileHandle);
-			fwrite($this->_fileHandle, serialize($this->_currentObject));
-			$this->_cellCache[$this->_currentObjectID]	= array('ptr' => $offset,
-																'sz'  => ftell($this->_fileHandle) - $offset
-															   );
+                fseek($this->_fileHandle,0,SEEK_END);
+                $offset = ftell($this->_fileHandle);
+                fwrite($this->_fileHandle, serialize($this->_currentObject));
+                $this->_cellCache[$this->_currentObjectID]	= array('ptr' => $offset,
+                                                                    'sz'  => ftell($this->_fileHandle) - $offset
+                                                                   );
+            }
 			$this->_currentCellIsDirty = false;
 		}
 		$this->_currentObjectID = $this->_currentObject = null;
